+++ udata_front/theme/gouvfr/datagouv-components/src/composables/useAccordion.ts
@@ -1,0 +1,17 @@
+import { ref } from "vue";
+import { toggleAccordion } from "../helpers";
+
+export default function useAccordion(id: string) {
+  const expanded = ref(false);
+  function expand() {
+    expanded.value = !expanded.value;
+    const element = document.getElementById(id);
+    if(element) {
+      toggleAccordion(element, expanded.value);
+    }
+  };
+  return {
+    expanded,
+    expand,
+  };
+}