+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 - Fix markdown tables and use DSFR ones [#432](https://github.com/datagouv/udata-front/pull/432)
 - Add organization selector to dataset publishing [#438](https://github.com/datagouv/udata-front/pull/438)
 - Add beta admin "add member" modale [#442](https://github.com/etalab/udata-front/pull/442)
-- Update dependencies from udata [#448](https://github.com/datagouv/udata-front/pull/448)
+- Update dependencies from udata [#448](https://github.com/datagouv/udata-front/pull/448) [#454](https://github.com/datagouv/udata-front/pull/454)
 - Add new dataset card [#445](https://github.com/datagouv/udata-front/pull/445)
+- Fix tests for oembed CORS [#453](https://github.com/datagouv/udata-front/pull/453)
 
 ## 5.0.1 (2024-06-12)